refactor: move all UI elements to the same level
