+++ src/app/join-us/JoinUsBox.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Link from 'next/link';
 import Image from 'next/image';
-import Button from '@/components/ui/Button/Button';
+import Button from '@/components/ui/Button';
 
 interface JoinUsBoxProps {
 	role: string;
