+++ src/components/profile/trainings.tsx
@@ -1,4 +1,4 @@
-import Button from '@/components/ui/Button/Button';
+import Button from '@/components/ui/Button';
 
 const Trainings = () => {
 	return (
